2:14AM
19/09/2024

+++ merchandiseMgmtERPSpringBoot/src/main/java/com/merchandisemgmt/merchandiseMgmtERP/entity/InventoryItem.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Entity
 @Data
 @NoArgsConstructor
@@ -18,7 +16,7 @@ public class InventoryItem {
     private Long id;
 
     @Column(nullable = false)
-    private Integer stock;
+    private long stock;
 
     @ManyToOne
     @JoinColumn
